requirements.txt
ccxt>=4.0.0
PyYAML>=6.0
pytest>=7.0.0
